requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.115.0
filelock==3.24.3
fsspec==2026.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
pillow==10.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
setuptools==82.0.0
starlette==0.38.6
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
